+++ src/token.zig
@@ -86,7 +86,7 @@ pub const Type = union(Types) {
     // Literals.
     IDENTIFIER: []const u8,
     STRING: []const u8,
-    NUMBER: f64,
+    NUMBER: i32,
 
     // Keywords.
     AND,
